Update spec to use offset and maxResults
Instead of using pageSize and pageNumber, use offset and maxResults
as this is a lot more flexible for API customers to get the data they
want (this may not seem obvious but there are scenarios in which the
user cannot get the data they want using a pageSize and pageNumber).

Simply put pageSize and pageNumber are tightly coupled while maxResults
and offset are two independant values (not coupled).