Merge version document into instance doc
This includes updating all methods for version endpoints to use the
instance store to update/get a version; reason for doing this is that
both an instance and a version are inherently the same thing and this
allows reduction of copying data across multiple documents.

Also use pointers in model so that nested objects are NOT returned if
the objects are empty.