Remove parallelisation of web endpoint unit tests
Tests were not guaranteed to pass due to variable names being the same
in each test that were being run in parallel. When accessing one of
these variables the variable may have changed due to a parallel test.

Now the tests always pass, the downside is an increase in the time taken
to run tests by an increase percentage of approx. 28% (from 0.035s to
0..045s).