Use new mock auditor from dependency go-ns
Previously audit parameters were being overwritten every time a new
audit message is created due to the underlying type of field is a map
and is inherently a pointer. New moack auditor uses a copy function to
copy parameters to a new pointer referenced in memory. This allows
audit parameters to be accurately tested in unit tests.