Prevent error from being masked by generic errs.ErrInternalServer
The default case sets err to the generic errs.ErrInternalServer err, which masks the original error detail and prevents it from being logged. The error is set to prevent error details from being leaked into the HTTP response, but in doing so it prevents the original error data from being logged.