Fix returning of struct
During local integration testing with race detect operating (which changed the system dynamic) on occasions it was observed that the mongo updates were not showing state of files changing from private to public.
The problem was down to returning the address of a struct passed on the stack.
With the change, integration tests now run OK.