Add basic local swagger spec linting
Added swagger spec linting using redocly. This has only been added for
local use and not added to the CI job at this point as it requires node
to be installed resulting in a combined go and node docker container
being needed. This can be added to CI as a later change.