Improve descriptions and examples
Improve the descriptions by fixing typos, clarifying terminology and
improving the clarity of language where possible for all public
endpoints that will be used for static datasets.

Add examples for all fields that will be publically used for static
datasets to better illustrate to a user what is expected. Also add
`format` attributes to date time string fields for clarity and to ensure
an accurate example is rendered.