Update spec to return latest version for edition
Update the API spec to merge the version model down into the edtions
model such that the edition is simply the latest version of that
edition. This partially addresses one of the main API user complaints
whereby they are required to determine the latest version every time in
order to get the distribution download URLs.

As the edition model is so simple, this can be achieved while
maintaining backwards compatability.

Further work is needed to understand whether the user complaint can be
fully addressed by returning the latest version of the latest edition at
the dataset series level (i.e. `/v1/datasets/{id}`) to prevent the user
from having to make that extra request when querying datasets that have
multiple editions and especially those that frequently release new
editions.