Ensure deprecated `downloads` are read only
The version `downloads` object should be returned by the versions
endpoints for backwards compatability, but as the field is deprecated
and the POST endpoint is new it should not form part of the request
body.

The existing PUT version endpoint uses a separate UpdateVersion model so
this change to the spec will not impact the implemenation of that
endpoint.