Stop returning database ID in response
The DB UUID for the version should not be returned in any of the API
responses as this should be internal to the dataset API only. This means
that the responses from the version write endpoints should be the same
model as the GET version endpoint.

The allOf with the collection ID is also redundant as the collection ID
is already in the version model.