Add `edition_title` and factor out id fields
Add the new `edition_title` field to provide a human readable edition
title for display on the UI.

Update the edition example values to better illustrate the difference
between the `edition` and `edition_title` fields.

Factor out the dataset and edition ids to improve consistency. Also add
the validation rules aligned to the URL standards as these ids are used
as part of the URLs.