Align Version and UpdateVersion models
Ensure the two version models are aligned by adding the missing fields
to the UpdateVersion model and ensuring the correct fields are marked as
readOnly in the Version model.

For the deprecated `downloads` map, the new create version endpoint
should not accept it in order to prevent its further use. However, as
the current update version endpoint allows these downloads to be
updated, the field should not be readOnly for this endpoint. This does
create some slight inconsistency and makes the modelling more complex,
but will ensure we are moving ahead with the deprecation.