Reduce duplication by reusing Version model
The spec had duplicate Version and UpdateVersion models that were
frequently becoming out of step from one another. The duplicate models
were redundant since marking the fields with readOnly appropriately can
achieve the same result.

The only complication is the fact that the update version endpoint
currently allows the `downloads` map to be updated for backwards
compatability where the create version endpoint does not since the field
is deprecated. This does result in a slightly ugly override of the
readOnly value for this field in the update version request body, but
this is preferrable to the duplication and drift experienced previously.