Replace redundant MetadataUpdate model
Replace usage of the redundant MetadataUpdate model with the Metadata
model now that appropriate readOnly attributes have been added.

This reduces the drift betweeen the two models for common fields. As the
code shares a common model for the editable fields this improves the
accuracy and reduces the maintenance burden of keeping the docs up to
date.