Remove unused fields
Remove the unused temporal and spatial coverage fields. This are
features needed by end users, but will be re-added in future when an
appropriate model can be agreed that better aligns to DCAT and common
language used by users. These fields are not exposed in the UI or
populated by the import processes so they can not be set currently. A
full scrape of all dataset records from the API also shows that they
have never been set.

Also remove the unused latest change type field as it is redundant and
has never been used. The admin UIs do not expose this field so it can
not be set currently. A full scrape of the dataset APIs shows no usages
of this field.