Reduce CI job durations
Reduce the CI job durations to improve team efficiency.

Introduce caching of pulled go modules to prevent the need to get them
on subsequent builds. This will only apply to CI builds run on the same
instance so will only improve the times for builds run a server where
the same task was run previously recently enough that the cache is still
populated.